+++ src/app/lhc-item-choice-check-box/lhc-item-choice-check-box.component.ts
@@ -56,7 +56,7 @@ export class LhcItemChoiceCheckBoxComponent implements OnInit, OnChanges {
 
       this.prevCheckBoxValue = this.item.value;
 
-      this.RemoveSubGroupsForNonExistentCheckboxes();
+      this.removeSubGroupsForNonExistentCheckboxes();
       this.updateSubGroupsForMergedQR();
     }
   }
@@ -97,12 +97,12 @@ export class LhcItemChoiceCheckBoxComponent implements OnInit, OnChanges {
     }
     for (let i = 0, len = this.checkboxModels.length; i < len; i++) {
       const checkboxOption = this.acOptions.listItems[i];
-      const subGroupLinkId = 'checkbox-subgroup-' + (checkboxOption.code || checkboxOption.text || checkboxOption.toString());
+      const subGroupLinkId = this.lhcDataService.getLhcFormData()._getLinkIdForCheckboxSubGroup(checkboxOption);
       const subGroupExists = this.lhcDataService.hasSubGroupWithLinkId(this.item, subGroupLinkId);
       if (this.checkboxModels[i] === true && !subGroupExists) {
-        this.lhcDataService.getLhcFormData().addSubItemsForCheckbox(this.item, checkboxOption, subGroupLinkId);
+        this.lhcDataService.getLhcFormData().addSubItemsForCheckbox(this.item, checkboxOption);
       } else if (!this.checkboxModels[i] && subGroupExists) {
-        this.lhcDataService.getLhcFormData().deleteSubItemsForCheckbox(this.item, checkboxOption._displayText, subGroupLinkId);
+        this.lhcDataService.getLhcFormData().deleteSubItemsForCheckbox(this.item, checkboxOption);
       }
     }
   }
@@ -132,7 +132,7 @@ export class LhcItemChoiceCheckBoxComponent implements OnInit, OnChanges {
    * removed from the answer options, due to answerExpression. The subgroups for those
    * removed checkboxes should also be removed.
    */
-  RemoveSubGroupsForNonExistentCheckboxes(): void {
+  removeSubGroupsForNonExistentCheckboxes(): void {
     // A list of currently valid subgroup linkIds.
     const subGroupLinkIds = this.acOptions.listItems.map(x =>
       'checkbox-subgroup-' + (x.code || x.text || x.toString()));